Move workflow scripts into respective entrypoint files
