+++ geograypher/entrypoints/__init__.py
@@ -1,0 +1,2 @@
+from geograypher.entrypoints.aggregate_images import aggregate_images
+from geograypher.entrypoints.render_labels import render_labels
