+++ geograypher/entrypoints/aggregate_images.py
@@ -1,14 +1,171 @@
 import argparse
+import typing
 from pathlib import Path
 
+import numpy as np
+
+from geograypher.cameras import MetashapeCameraSet
 from geograypher.constants import (
     EXAMPLE_CAMERAS_FILENAME,
     EXAMPLE_IDS_TO_LABELS,
     EXAMPLE_IMAGE_FOLDER,
     EXAMPLE_MESH_FILENAME,
     EXAMPLE_PREDICTED_LABELS_FOLDER,
+    PATH_TYPE,
 )
 from geograypher.entrypoints.workflow_functions import aggregate_images
+from geograypher.meshes import TexturedPhotogrammetryMesh
+from geograypher.segmentation import SegmentorPhotogrammetryCameraSet
+from geograypher.segmentation.derived_segmentors import LookUpSegmentor
+
+
+def aggregate_images(
+    mesh_file: PATH_TYPE,
+    cameras_file: PATH_TYPE,
+    image_folder: PATH_TYPE,
+    label_folder: PATH_TYPE,
+    subset_images_folder: typing.Union[PATH_TYPE, None] = None,
+    mesh_transform_file: typing.Union[PATH_TYPE, None] = None,
+    DTM_file: typing.Union[PATH_TYPE, None] = None,
+    height_above_ground_threshold: float = 2.0,
+    ROI: typing.Union[PATH_TYPE, None] = None,
+    ROI_buffer_radius_meters: float = 50,
+    IDs_to_labels: typing.Union[dict, None] = None,
+    mesh_downsample: float = 1.0,
+    aggregate_image_scale: float = 1.0,
+    aggregated_face_values_savefile: typing.Union[PATH_TYPE, None] = None,
+    predicted_face_classes_savefile: typing.Union[PATH_TYPE, None] = None,
+    top_down_vector_projection_savefile: typing.Union[PATH_TYPE, None] = None,
+    vis: bool = False,
+):
+    """Aggregate labels from multiple viewpoints onto the surface of the mesh
+
+    Args:
+        mesh_file (PATH_TYPE):
+            Path to the Metashape-exported mesh file
+        cameras_file (PATH_TYPE):
+            Path to the MetaShape-exported .xml cameras file
+        image_folder (PATH_TYPE):
+            Path to the folder of images used to create the mesh
+        label_folder (PATH_TYPE):
+            Path to the folder of labels to be aggregated onto the mesh. Must be in the same
+            structure as the images
+        subset_images_folder (typing.Union[PATH_TYPE, None], optional):
+            Use only images from this subset. Defaults to None.
+        mesh_transform_file (typing.Union[PATH_TYPE, None], optional):
+            Transform from the mesh coordinates to the earth-centered, earth-fixed frame. Can be a
+            4x4 matrix represented as a .csv, or a Metashape cameras file containing the
+            information. Defaults to None.
+        DTM_file (typing.Union[PATH_TYPE, None], optional):
+            Path to a digital terrain model file to remove ground points. Defaults to None.
+        height_above_ground_threshold (float, optional):
+            Height in meters above the DTM to consider ground. Only used if DTM_file is set.
+            Defaults to 2.0.
+        ROI (typing.Union[PATH_TYPE, None], optional):
+            Geofile region of interest to crop the mesh to. Defaults to None.
+        ROI_buffer_radius_meters (float, optional):
+            Keep points within this distance of the provided ROI object, if unset, everything will
+            be kept. Defaults to 50.
+        IDs_to_labels (typing.Union[dict, None], optional):
+            Maps from integer IDs to human-readable class name labels. Defaults to None.
+        mesh_downsample (float, optional):
+            Downsample the mesh to this fraction of vertices for increased performance but lower
+            quality. Defaults to 1.0.
+        aggregate_image_scale (float, optional):
+            Downsample the labels before aggregation for faster runtime but lower quality. Defaults
+            to 1.0.
+        aggregated_face_values_savefile (typing.Union[PATH_TYPE, None], optional):
+            Where to save the aggregated image values as a numpy array. Defaults to None.
+        predicted_face_classes_savefile (typing.Union[PATH_TYPE, None], optional):
+            Where to save the most common label per face texture as a numpy array. Defaults to None.
+        top_down_vector_projection_savefile (typing.Union[PATH_TYPE, None], optional):
+            Where to export the predicted map. Defaults to None.
+        vis (bool, optional):
+            Show the mesh model and predicted results. Defaults to False.
+    """
+    ## Create the camera set
+    # Do the camera operations first because they are fast and good initial error checking
+    camera_set = MetashapeCameraSet(cameras_file, image_folder)
+
+    # If the ROI is not None, subset to cameras within a buffer distance of the ROI
+    # TODO let get_subset_ROI accept a None ROI and return the full camera set
+    if subset_images_folder is not None:
+        camera_set = camera_set.get_cameras_in_folder(subset_images_folder)
+
+    if ROI is not None and ROI_buffer_radius_meters is not None:
+        # Extract cameras near the training data
+        camera_set = camera_set.get_subset_ROI(
+            ROI=ROI, buffer_radius_meters=ROI_buffer_radius_meters
+        )
+
+    if mesh_transform_file is None:
+        mesh_transform_file = cameras_file
+
+    ## Create the mesh
+    mesh = TexturedPhotogrammetryMesh(
+        mesh_file,
+        transform_filename=mesh_transform_file,
+        ROI=ROI,
+        ROI_buffer_meters=ROI_buffer_radius_meters,
+        IDs_to_labels=IDs_to_labels,
+        downsample_target=mesh_downsample,
+    )
+
+    # Show the mesh if requested
+    if vis:
+        mesh.vis(camera_set=camera_set)
+
+    # Create a segmentor object to load in the predictions
+    segmentor = LookUpSegmentor(
+        base_folder=image_folder,
+        lookup_folder=label_folder,
+        num_classes=np.max(list(mesh.get_IDs_to_labels().keys())) + 1,
+    )
+    # Create a camera set that returns the segmented images instead of the original ones
+    segmentor_camera_set = SegmentorPhotogrammetryCameraSet(
+        camera_set, segmentor=segmentor
+    )
+
+    ## Perform aggregation
+    # this is the slow step
+    aggregated_face_labels, _, _ = mesh.aggregate_viewpoints_pytorch3d(
+        segmentor_camera_set,
+        image_scale=aggregate_image_scale,
+    )
+    # If requested, save this data
+    if aggregated_face_values_savefile is not None:
+        Path(aggregated_face_values_savefile).parent.mkdir(exist_ok=True, parents=True)
+        np.save(aggregated_face_values_savefile, aggregated_face_labels)
+
+    # Find the most common class per face
+    predicted_face_classes = np.argmax(
+        aggregated_face_labels, axis=1, keepdims=True
+    ).astype(float)
+
+    # If requested, label the ground faces
+    if DTM_file is not None and height_above_ground_threshold is not None:
+        predicted_face_classes = mesh.label_ground_class(
+            labels=predicted_face_classes,
+            height_above_ground_threshold=height_above_ground_threshold,
+            DTM_file=DTM_file,
+            ground_ID=np.nan,
+            set_mesh_texture=False,
+        )
+
+    if predicted_face_classes_savefile is not None:
+        Path(predicted_face_classes_savefile).parent.mkdir(exist_ok=True, parents=True)
+        np.save(predicted_face_classes_savefile, predicted_face_classes)
+
+    if vis:
+        # Show the mesh with predicted classes
+        mesh.vis(vis_scalars=predicted_face_classes)
+
+    # TODO this should be updated to take IDs_to_labels
+    mesh.export_face_labels_vector(
+        face_labels=np.squeeze(predicted_face_classes),
+        export_file=top_down_vector_projection_savefile,
+        vis=True,
+    )
 
 
 def parse_args():
